Refactor to shorten _listRelationshipField function
