+++ src/packages/core/src/utils/common.ts
@@ -1,0 +1,2 @@
+export const isDefined = <T>(value: T | undefined | null): value is T =>
+	value !== undefined && value !== null;
