+++ src/packages/core/src/utils/index.ts
@@ -2,3 +2,4 @@ export * from './apply-default-values';
 export * from './has-id';
 export * from './plural';
 export * from './with-transaction';
+export * from './common';
